Simplify replication controller
